Switch return to long
According to miod@, syscall actually returns a "register_t" that
is an old BSDism and so instead we should use a "long".

"long is better than int in all cases"

This will allow us to return pointers from the syscalls that return them.

+++ files/syscall_emulator.h
@@ -1,1 +1,1 @@
-int	syscall_emulator(int, ...);
+long	syscall_emulator(int, ...);
